# backend/docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  handlers.CreateCreditRequestRequest:
    description: Datos para crear una nueva solicitud de crédito
    properties:
      amount:
        example: 10000000
        type: number
      creditStatusId:
        example: 1
        type: integer
      customerId:
        example: 1
        type: integer
      productType:
        example: Préstamo Personal
        type: string
      termMonths:
        example: 24
        type: integer
    type: object
...
  handlers.UpdateCreditRequestRequest:
    description: Datos para actualizar una solicitud de crédito existente
    properties:
      amount:
        example: 15000000
        type: number
      creditStatusId:
        example: 2
        type: integer
      customerId:
        example: 1
        type: integer
      productType:
        example: Hipoteca
        type: string
      termMonths:
        example: 36
        type: integer
    type: object
...
  models.CreditRequest:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      creditStatusId:
        type: integer
      customerId:
        type: integer
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      productType:
        type: string
      riskCategory:
        type: string
      riskExplanation:
        type: string
      riskScore:
        type: number
      termMonths:
        type: integer
      updatedAt:
        type: string
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@example.com
    name: API Support
  description: API para gestión de solicitudes de crédito y evaluación de riesgo crediticio.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Credit Risk Assessment API
  version: "1.0"
paths:
...
  /credit-requests:
    get:
      consumes:
      - application/json
      description: Retorna una lista de todas las solicitudes de crédito, opcionalmente
        filtradas por cliente
      parameters:
      - description: ID del cliente para filtrar
        in: query
        name: customerId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de solicitudes de crédito
          schema:
            items:
              $ref: '#/definitions/models.CreditRequest'
            type: array
        "400":
          description: customerId inválido
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Obtener todas las solicitudes de crédito
      tags:
      - Credit Requests
    post:
      consumes:
      - application/json
      description: Crea una nueva solicitud de crédito en el sistema
      parameters:
      - description: Datos de la solicitud de crédito
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCreditRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Solicitud de crédito creada exitosamente
          schema:
            $ref: '#/definitions/models.CreditRequest'
        "400":
          description: Solicitud inválida
          schema:
            type: string
        "404":
          description: Cliente o estado de crédito no encontrado
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Crear una nueva solicitud de crédito
      tags:
      - Credit Requests
  /credit-requests/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina una solicitud de crédito del sistema
      parameters:
      - description: ID de la solicitud de crédito
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Solicitud eliminada exitosamente
        "400":
          description: ID inválido
          schema:
            type: string
        "404":
          description: Solicitud no encontrada
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Eliminar una solicitud de crédito
      tags:
      - Credit Requests
    get:
      consumes:
      - application/json
      description: Retorna los detalles de una solicitud de crédito específica
      parameters:
      - description: ID de la solicitud de crédito
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Solicitud de crédito encontrada
          schema:
            $ref: '#/definitions/models.CreditRequest'
        "400":
          description: ID inválido
          schema:
            type: string
        "404":
          description: Solicitud no encontrada
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Obtener una solicitud de crédito por ID
      tags:
      - Credit Requests
    put:
      consumes:
      - application/json
      description: Actualiza los datos de una solicitud de crédito existente
      parameters:
      - description: ID de la solicitud de crédito
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados de la solicitud
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateCreditRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Solicitud actualizada exitosamente
          schema:
            $ref: '#/definitions/models.CreditRequest'
        "400":
          description: Solicitud inválida
          schema:
            type: string
        "404":
          description: Solicitud no encontrada
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Actualizar una solicitud de crédito
      tags:
      - Credit Requests
...
securityDefinitions:
  BearerAuth:
    description: 'Ingresa el token JWT con el prefijo Bearer. Ejemplo: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"